Add log viewer
On the Window menu there is now a "Show Log" item which opens a window
displaying a log of recent actions. Currently, only actions involving an
action window are logged. The action is logged when you request the action
(so it might not have actually been performed).

The log window was created using Glade, and this patch also adds a Glade
dependency on ROX-Filer, along with some support code to use it.