Added preliminary support for Glade translations (as suggested by Lennon Cook).
We check for the presence of libintl.h and use the bindtextdomain() call
to tell Glade where to find its catalogs. Right now the $APP_DIR/Messages
directory must be set up manually so that the catalogs can be found, eg.:

	.
	|-- hu
	|   `-- LC_MESSAGES
	|       `-- ROX-Filer.mo -> ../../hu.gmo
	`-- hu.gmo

This seems to work fine when ROX-Filer is configured to use the LANG
environment variable. Changing languages manually is not supported
though, since we should change the locale on the fly, which does not
seem to be a good idea to me...