Fixed check to ensure that the ROX theme is available.
Before, we just checked that the postscript icon could be found, but
now we fallback to the GNOME theme so that always succeeds. Also, if
the theme exists but is a broken link then we fix it automatically
rather than popping up a box telling the user to fix it.