Removed separate source feed
For historical reasons, ROX-Filer had a somewhat confused build system. The
main feed claimed it to be a binary that worked on any architecture (because it
compiled itself if the binary was missing), while a separate "source" feed
provided the build dependencies.

Now it has just a single feed, asserting that this is source code. Compiling it
generates a binary release, including a binary feed.

Also, there is a new 0compile.properties file so you can easily compile it with
"0compile build".

Clicking on the source's AppRun when there is no binary compiles it as before
(without using Zero Install).