Allow sorting by atime and ctime as well as mtime.
In addition to the new sort types, this also adds atime and ctime
columns to the list view -- which makes sense, given that the details
view shows atime/ctime when times are shown. I've kept mtime first
because it's the one you're most likely to be interested in.

For consistency, it would make sense to rename "Sort by Date" to "Sort
by Last Modified" (and SORT_DATE to SORT_MTIME), but that would need
changes to all the translations too.